requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
face-recognition==1.3.0
opencv-python==4.9.0.80
numpy==1.26.4
pillow==10.2.0
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
redis==5.0.1
